Created Pistol Blueprint
Created the Pistol Blueprint as a child class of BaseWeapon.  Set PistolIdle as the Source Flipbook for the PaperFlipbook component.  Set the GunDamage value to be equal to 1.  Implemented the Pistol's EventGraph to decrement the PistolAmmo variable and to cycle between the PistolIdle and PistolFire Flipbooks whenever the BaseWeapon's Fire event is called.