Weapon ChildActor
Added the Weapon component as a ChildActor of the FirstPersonCamera.   Set the Child Actor Class to be the Pistol. Adjusted the Pistol's PaperFlipbook Scale values to be 0.1, and set the FirstPersonCharacter's PaperFlipbook Location values so that X = 42 and Z = -21.  Implemented the FirstPersonCharacter's Spawn Projectile section in the Event Graph.  Modified the BaseWeapon's Fire event so that the LineTraceForObjects' Draw Debug Type equals "For Duration," so that the Line Trace can be visualized after firing the Pistol.  Adjusted the Pistol's WeaponTrace Y Location to be -670, so that the LineTrace begins in front of the Weapon ChildActor.