Created Shotgun
Added float variable ShotgunAmmo to the BaseWeapon Blueprint and initialized its Default Value to 20.  Created the Shotgun Blueprint as a Child Blueprint of BaseWeapon.  Set the Shotgun's GunDamage value to 2.  Added an Arrow component to the Shotgun's Default Scene Root and adjusted the Scaling / Location values for the Shotgun's PaperFlipbook.  Implemented the Event Attack logic in the Event Graph.