Created Minigun
Modified the BaseWeapon Blueprint to include an Arrow component, then adjusted the rotation / scaling values for the PaperFlipbook and WeaponTrace components.   Created the Minigun Blueprint as a child of BaseWeapon.  Created the Minigun sprites using Paper2D, then generated Flipbooks for MinigunFire and MinigunIdle.  Set the PaperFlipbook's Source Flipbook value to MinigunIdle.  Set the GunDamage value to 1, and set its ammo source to PistolAmmo.  Implemented the Event Graph, including the scripts for Event Attack and Event AttackStop.