Created RpgProjectile
Imported the resources needed to create the RpgProjectile Blueprint.  Implemented the RpgProjectile EventGraph so that the projectile moves for every Tick.  Set gravity values to 0. Used Paper2D to generate  sprites needed for the RpgExplosion Flipbook.  Created the RpgExplosionActor and implemented the EventGraph so that the explosion animation is always facing the player (yaw rotation).  Finally, set up the RpgProjectile's EventGraph to call upon the RpgExplosionActor, which plays a PaperFlipbook animation whenever collision is detected against World Static and Pawn objects.