Created HealthPickup
Imported the necessary resources for creating the pickup sprites via Paper2D.  Created the HealthPickup Blueprint class and modified its Event Graph so that it now includes node networks for the pickup's box collision, as well as Event Tick yaw rotations to ensure that the HealthPickup sprite is always player-facing.  Included a Clamp Float node so that the player's Health value cannot exceed 100, and created branching so that the HealthPickup Actor will only be destroyed if the player's health is less than 100 at the time of collision.