Created Ammo Pickups
Created Blueprint classes for the PistolAmmoPickup, ShotgunAmmoPickup, and RpgAmmoPickup.  Modified each pickup's Event Graph to implement networks for handling box collision, as well as Event Tick yaw rotations to ensure that the pickup sprites are always player-facing.  Included Clamp Float nodes so that the player's Pistol, Shotgun, and RPG ammo values cannot exceed 50, 30, and 10.  Created branching so that the pickup's Actor will only be destroyed if the player's ammo value is less than the max Clamp Float value at the time of collision.