MeleeEnemy FlipbookState
Created the MeleeEnemy_FlipbookState Blueprint Enumeration and assigned it to the MeleeEnemy's FlipbookState variable.  Also created PaperFlipbook Object Reference Array variables for IdleState, RunState, and AttackState.  Implemented the MeleeEnemy's Event Graph to include node networks for setting the MeleeEnemy's PaperFlipbook sprite to either Front, Right, Left, or Back--depending on the player's perspective in relation to the MeleeEnemy.