RangedEnemy Weapon Projectile
Imported the RangedEnemy FireTexture, then used it to extract the FireSprites.  Used the FireSprites to generate the FireFlipbook.  Created the RangedEnemy_WeaponProjectile Actor Blueprint.  Added an Arrow component and a PaperFlipbook component with Sphere collision.  Set the radius of the Sphere collision to 10.  Added a ProjectileMovement component.  Modified the EventGraph to include node networks for Event Tick Yaw Rotations, (Sphere) On Component Begin Overlap, and DestroyActor Withuout Collision.