Implemented KeycardDoor
Modified the FirstPersonCharacter Blueprint to include the Boolean variable IsHoldingKeycard.  Imported the KeycardItem's StaticMesh and Texture resources.  Generated a StaticMesh from the Keycard Texture.  Created the KeycardItem Blueprint Actor.  Set its StaticMesh to the Keycard StaticMesh.  Added a Box Collision component and implemented the node network for OnComponentBeginOverlap.  Created the Widget for FindKeycard and implemented the node network for Event Construct.  Modified the KeycardDoor's Event Graph to now include a node network for the Box Collision's OnComponentBeginOverlap.