Implemented LavaPool
Imported the Acid and Lava textures.  Created the Lava Material using the Lava texture.  Created the Lava Blueprint Actor and gave it StaticMesh and Box Collision components.  Set the StaticMesh's Material to be the Lava Material.  Set scale for the Static Mesh and implemented the Box Collision's OnComponentBeginOverlap network.  Modified the FirstPersonCharacter's Player Damage system to include a custom function named InstantKill.  This function is called whenever the player overlaps the Lava's Box Collision component.  Modified the Lava Material to include a Panner node and a TexCoord[0] node, for controlling the texture's speed and scaling quality.