Implemented AcidPool
Created the AcidPool Material, as well as the AcidPool Blueprint Actor.  Modified the Blueprint's EventGraph so that instead of calling InstantKill on ComponentBeginOverlap, there's a call to toggle the value of the IsPlayerAcidPool variable.  Implemented a node network for for Event BeginPlay which includes a network for the custom event ApplyAcidDamage.