AcidPool Damage, Ammo Values, Keycard Frame
Modified the AcidPool Blueprint so that it now applies even greater damage at a slightly faster rate.  Modified the initial ammo values so that the player starts with nothing.  Used BSP brushes to create the KeycardFrame and then placed it into the AcidMap level after converting the BSPs into a single Static Mesh.